Fixed memory leaks
When rendering the visual boxes of the collision boxes, the method was constructing the geometry each time it was called. This was causing memory leaks because I think the CBox class uses dynamic memory allocation which is not freed. This was easily fixed with the static objects as the geometry only needs to be constructed at the beginning of the program, but for the carriages I had to just take out the line.